+++ pkg/irsa/capa/capa.go
@@ -22,6 +22,7 @@ import (
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/cloudfront"
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/iam"
 	"github.com/giantswarm/irsa-operator/pkg/aws/services/s3"
+	"github.com/giantswarm/irsa-operator/pkg/errors"
 	"github.com/giantswarm/irsa-operator/pkg/key"
 	ctrlmetrics "github.com/giantswarm/irsa-operator/pkg/metrics"
 	"github.com/giantswarm/irsa-operator/pkg/util"
@@ -52,7 +53,10 @@ func (s *Service) Reconcile(ctx context.Context) error {
 
 	s.Scope.Info("Reconciling AWSCluster CR for IRSA")
 	privateKey, err := s.ServiceAccountSecret(ctx)
-	if err != nil {
+	if apierrors.IsNotFound(err) {
+		s.Scope.Info("Service account is not ready yet, waiting ...")
+		return nil
+	} else if err != nil {
 		return err
 	}
 
@@ -118,6 +122,12 @@ func (s *Service) Reconcile(ctx context.Context) error {
 	cfConfig := &v1.Secret{}
 	err = s.Client.Get(ctx, types.NamespacedName{Namespace: s.Scope.ClusterNamespace(), Name: s.Scope.ConfigName()}, cfConfig)
 	if apierrors.IsNotFound(err) {
+		if err := errors.IsEmptyCloudfrontDistribution(distribution); err != nil {
+			ctrlmetrics.Errors.WithLabelValues(s.Scope.Installation(), s.Scope.AccountID(), s.Scope.ClusterName(), s.Scope.ClusterNamespace()).Inc()
+			s.Scope.Logger.Error(err, "cloudfront distribution cannot be nil")
+			return err
+		}
+
 		// create new OIDC Cloudfront config
 		cfConfig := &v1.Secret{
 			ObjectMeta: metav1.ObjectMeta{
@@ -313,7 +323,6 @@ func (s *Service) ServiceAccountSecret(ctx context.Context) (*rsa.PrivateKey, er
 	oidcSecret := &v1.Secret{}
 	err := s.Client.Get(ctx, types.NamespacedName{Namespace: s.Scope.ClusterNamespace(), Name: s.Scope.ClusterName() + "-sa"}, oidcSecret)
 	if err != nil {
-		s.Scope.Logger.Error(err, "failed to get service account secret")
 		return nil, err
 	}
 	privBytes := oidcSecret.Data["tls.key"]