+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Check if Cloudfront Distribution is empty.
+
 ## [0.8.3] - 2022-10-28
 
 ### Fixed